Moved core files to pysumma/, got rid of OLD test files
