+++ pysumma/Simulation.py
@@ -1,6 +1,6 @@
 #from Decisions import Decisions         # This is for testing in cmd window.
-from pysumma.Option import Option
-from .Decisions import Decisions       # This is for testing in this python code.
+from Option import Option
+from Decisions import Decisions       # This is for testing in this python code.
 import subprocess
 import os
 import xarray as xr
